tools/eval-orchestrator/pyproject.toml
[project]
name = "eval-orchestrator"
version = "0.1.0"
description = "Split tolokaforge run configs into shards for parallel CI execution, and merge results"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.5,<0.13",
    "click>=8.0.0,<8.2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
eval-orchestrator = "eval_orchestrator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eval_orchestrator"]
